I saw my mistake as soon as I started...
The idea was to have the manager handle a container of
abstract settings, but now I realized this won't work with
the acstract class being a template, because without specifying
the type for the template, it's not really a thing...

So I have to not use a template after all... Ho hum.

I guess I need these types of settings:

+ String [done]
+ long signed integer
+ byte (can act as a bool with minimal space)
+ float

So I will need to de-templatize these... boo.