Now working!
So we have basic DHCP mode working - we can connect to the AP, set
the wifi credentials and connect. This is awesome. :)

All this with a smaller footprint than WiFi Manager library, and
works in a non-blocking way so we can still have our code doing
something else while switching around network stuff (potentially - I
need to work out how this is to be done.

TODO:

1.  Add custom parameters page
2.  Enable setting of host name
3.  Add host name to "connecting" after setting wifi
3.  Re-design front page a bit
4.  Show "no networks found" message and add refresh button to Wifi page